lots of deduping of types and consolidation of files
