+++ src/entities/transactions/index.ts
@@ -1,7 +1,4 @@
-export type { NewTransaction, NewTransactionOrAddCashTransaction, RainbowTransaction } from './transaction';
-export { default as TransactionStatusTypes, TransactionStatus } from './transactionStatus';
-export { transactionTypes } from './transactionType';
+export * from './transaction';
 
 export type { ZerionTransaction, ZerionTransactionChange } from './zerionTransaction';
-export { default as TransactionDirections, TransactionDirection } from './transactionDirection';
 export { default as GasFeeTypes, GasFeeType } from './gasFeeType';
