--- src/entities/transactions/transactionDirection.ts
@@ -1,11 +1,0 @@
-export enum TransactionDirection {
-  in = 'in',
-  out = 'out',
-  self = 'self',
-}
-
-export default {
-  in: 'in',
-  out: 'out',
-  self: 'self',
-};
