--- src/entities/transactions/transactionStatus.ts
@@ -1,61 +1,0 @@
-export enum TransactionStatus {
-  approved = 'approved',
-  approving = 'approving',
-  bridging = 'bridging',
-  bridged = 'bridged',
-  cancelled = 'cancelled',
-  cancelling = 'cancelling',
-  contract_interaction = 'contract interaction',
-  deposited = 'deposited',
-  depositing = 'depositing',
-  dropped = 'dropped',
-  failed = 'failed',
-  minted = 'minted',
-  minting = 'minting',
-  purchased = 'purchased',
-  purchasing = 'purchasing',
-  received = 'received',
-  receiving = 'receiving',
-  self = 'self',
-  selling = 'selling',
-  sold = 'sold',
-  sending = 'sending',
-  sent = 'sent',
-  speeding_up = 'speeding up',
-  swapped = 'swapped',
-  swapping = 'swapping',
-  unknown = 'unknown status',
-  withdrawing = 'withdrawing',
-  withdrew = 'withdrew',
-}
-
-export default {
-  approved: 'approved',
-  approving: 'approving',
-  bridging: 'bridging',
-  bridged: 'bridged',
-  cancelled: 'cancelled',
-  cancelling: 'cancelling',
-  contract_interaction: 'contract interaction',
-  deposited: 'deposited',
-  depositing: 'depositing',
-  dropped: 'dropped',
-  failed: 'failed',
-  minted: 'minted',
-  minting: 'minting',
-  purchased: 'purchased',
-  purchasing: 'purchasing',
-  received: 'received',
-  receiving: 'receiving',
-  self: 'self',
-  selling: 'selling',
-  sold: 'sold',
-  sending: 'sending',
-  sent: 'sent',
-  speeding_up: 'speeding up',
-  swapped: 'swapped',
-  swapping: 'swapping',
-  unknown: 'unknown status',
-  withdrawing: 'withdrawing',
-  withdrew: 'withdrew',
-};
