--- src/entities/transactions/transactionType.ts
@@ -1,24 +1,0 @@
-export const transactionTypes = {
-  withoutChanges: ['cancel', 'contract_interaction', 'deployment', 'approve', 'revoke', 'speed_up'],
-  withChanges: [
-    'sale',
-    'bridge',
-    'airdrop',
-    'wrap',
-    'unwrap',
-    'bid',
-    'burn',
-    'send',
-    'receive',
-    'withdraw',
-    'deposit',
-    'mint',
-    'swap',
-    'borrow',
-    'claim',
-    'repay',
-    'stake',
-    'unstake',
-    'purchase',
-  ],
-} as const;
