+++ src/parsers/transactions.ts
@@ -1,6 +1,14 @@
-import { slice } from 'lodash';
-import { parseAllTxnsOnReceive } from '../config/debug';
-import { NativeCurrencyKey, RainbowTransaction, ZerionTransaction } from '@/entities';
+import {
+  NativeCurrencyKey,
+  RainbowTransaction,
+  TransactionDirection,
+  PaginatedTransactionsApiResponse,
+  TransactionApiResponse,
+  TransactionChanges,
+  TransactionStatus,
+  TransactionType,
+  TransactionWithChangesType,
+} from '@/entities';
 
 import {
   convertAmountAndPriceToNativeDisplay,
@@ -12,17 +20,8 @@ import {
 import { NewTransaction, RainbowTransactionFee } from '@/entities/transactions/transaction';
 import { parseAddressAsset, parseAsset } from '@/resources/assets/assets';
 import { ParsedAsset } from '@/resources/assets/types';
-import { transactionTypes } from '@/entities/transactions/transactionType';
-import {
-  PaginatedTransactionsApiResponse,
-  TransactionApiResponse,
-  TransactionChanges,
-  TransactionType,
-  TransactionWithChangesType,
-} from '@/resources/transactions/types';
-import { ChainId } from '@/chains/types';
 
-const LAST_TXN_HASH_BUFFER = 20;
+import { ChainId } from '@/chains/types';
 
 const TransactionOutTypes = [
   'burn',
@@ -37,25 +36,11 @@ const TransactionOutTypes = [
   'revoke',
   'deployment',
   'contract_interaction',
-] as const;
+] as readonly string[];
 
-export const getDirection = (type: TransactionType) => {
-  // @ts-expect-error - Ts doesnt like the weird type structure here
-  if (TransactionOutTypes.includes(type as TransactionType)) return 'out';
-  return 'in';
-};
-
-const dataFromLastTxHash = (transactionData: ZerionTransaction[], transactions: RainbowTransaction[]): ZerionTransaction[] => {
-  if (__DEV__ && parseAllTxnsOnReceive) return transactionData;
-  const lastSuccessfulTxn = transactions.find(txn => !!txn.hash && txn.status !== 'pending');
-  const lastTxHash = lastSuccessfulTxn?.hash;
-  if (lastTxHash) {
-    const lastTxnHashIndex = transactionData.findIndex(txn => lastTxHash.startsWith(txn.hash));
-    if (lastTxnHashIndex > -1) {
-      return slice(transactionData, 0, lastTxnHashIndex + LAST_TXN_HASH_BUFFER);
-    }
-  }
-  return transactionData;
+export const getDirection = (type: TransactionType): TransactionDirection => {
+  if (TransactionOutTypes.includes(type)) return TransactionDirection.OUT;
+  return TransactionDirection.IN;
 };
 
 export const getAssetFromChanges = (changes: TransactionChanges, type: TransactionType) => {
@@ -147,12 +132,12 @@ export const parseTransaction = async (
   } as RainbowTransaction;
 };
 
-export const parseNewTransaction = (tx: NewTransaction): RainbowTransaction => {
+export const convertNewTransactionToRainbowTransaction = (tx: NewTransaction): RainbowTransaction => {
   const asset = tx?.changes?.[0]?.asset || tx.asset;
 
   return {
     ...tx,
-    status: 'pending',
+    status: TransactionStatus.pending,
     data: tx.data,
     title: `${tx.type}.${tx.status}`,
     description: asset?.name,
